Refactored `files` and Added `observeAlbums` and `observeArtists`
