+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/methods/queries/AllPathQuery.kt
@@ -1,10 +1,10 @@
-package com.lucasjosino.on_audio_query.query
+package com.lucasjosino.on_audio_query.methods
 
 import android.content.ContentResolver
 import android.content.Context
 import android.net.Uri
 import android.provider.MediaStore
-import com.lucasjosino.on_audio_query.controller.PermissionController
+import com.lucasjosino.on_audio_query.controllers.PermissionController
 import io.flutter.plugin.common.MethodChannel
 import java.io.File
