+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/methods/queries/ArtistsQuery.kt
@@ -1,23 +1,23 @@
-package com.lucasjosino.on_audio_query.query
+package com.lucasjosino.on_audio_query.methods
 
 import android.content.ContentResolver
 import android.content.Context
 import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.controller.PermissionController
-import com.lucasjosino.on_audio_query.query.helper.QueryHelper
-import com.lucasjosino.on_audio_query.types.checkPlaylistsUriType
-import com.lucasjosino.on_audio_query.types.sorttypes.checkGenreSortType
-import com.lucasjosino.on_audio_query.utils.playlistProjection
+import com.lucasjosino.on_audio_query.controllers.PermissionController
+import com.lucasjosino.on_audio_query.methods.helper.QueryHelper
+import com.lucasjosino.on_audio_query.types.checkArtistsUriType
+import com.lucasjosino.on_audio_query.types.sorttypes.checkArtistSortType
+import com.lucasjosino.on_audio_query.utils.artistProjection
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-/** PlaylistsQuery */
-class PlaylistsQuery : ViewModel() {
+/** ArtistsQuery */
+class ArtistsQuery : ViewModel() {
 
     //Main parameters
     private val helper = QueryHelper()
@@ -28,18 +28,18 @@ class PlaylistsQuery : ViewModel() {
     private lateinit var sortType: String
 
     /**
-     * Method to "query" all playlists.
+     * Method to "query" all artists.
      *
      * Parameters:
      *   * [context]
      *   * [result]
      *   * [call]
      */
-    fun queryPlaylists(context: Context, result: MethodChannel.Result, call: MethodCall) {
+    fun queryArtists(context: Context, result: MethodChannel.Result, call: MethodCall) {
         resolver = context.contentResolver
 
-        // Sort: Type and Order.
-        sortType = checkGenreSortType(
+        // Sort: Type and Order
+        sortType = checkArtistSortType(
             call.argument<Int>("sortType"),
             call.argument<Int>("orderType")!!,
             call.argument<Boolean>("ignoreCase")!!
@@ -47,63 +47,59 @@ class PlaylistsQuery : ViewModel() {
         // Check uri:
         //   * [0]: External.
         //   * [1]: Internal.
-        uri = checkPlaylistsUriType(call.argument<Int>("uri")!!)
+        uri = checkArtistsUriType(call.argument<Int>("uri")!!)
 
         // Query everything in background for a better performance.
         viewModelScope.launch {
             // Request permission status from the main method.
             val hasPermission = PermissionController().permissionStatus(context)
             // Empty list.
-            var resultPlaylistList = ArrayList<MutableMap<String, Any?>>()
+            var resultArtistList = ArrayList<MutableMap<String, Any?>>()
 
             // We cannot "query" without permission so, just return a empty list.
             if (hasPermission) {
                 // Start querying
-                resultPlaylistList = loadPlaylists()
+                resultArtistList = loadArtists()
             }
 
             //Flutter UI will start, but, information still loading
-            result.success(resultPlaylistList)
+            result.success(resultArtistList)
         }
     }
 
-    //Loading in Background
-    private suspend fun loadPlaylists(): ArrayList<MutableMap<String, Any?>> =
+    // Loading in Background
+    private suspend fun loadArtists(): ArrayList<MutableMap<String, Any?>> =
         withContext(Dispatchers.IO) {
-            // Setup the cursor with [uri] and [projection].
-            val cursor = resolver.query(uri, playlistProjection, null, null, null)
+            // Setup the cursor with [uri], [projection] and [sortType].
+            val cursor = resolver.query(uri, artistProjection, null, null, sortType)
             // Empty list.
-            val playlistList: ArrayList<MutableMap<String, Any?>> = ArrayList()
+            val artistList: ArrayList<MutableMap<String, Any?>> = ArrayList()
 
-            // For each item(playlist) inside this "cursor", take one and "format"
+            // For each item(artist) inside this "cursor", take one and "format"
             // into a [Map<String, dynamic>].
             while (cursor != null && cursor.moveToNext()) {
-                val playlistData: MutableMap<String, Any?> = HashMap()
-                for (playlistMedia in cursor.columnNames) {
-                    playlistData[playlistMedia] = helper.loadPlaylistItem(playlistMedia, cursor)
+                val tempData: MutableMap<String, Any?> = HashMap()
+                for (artistMedia in cursor.columnNames) {
+                    tempData[artistMedia] = helper.loadArtistItem(artistMedia, cursor)
                 }
 
-                // Count and add the number of songs for every playlist.
-                val mediaCount = helper.getMediaCount(1, playlistData["_id"].toString(), resolver)
-                playlistData["num_of_songs"] = mediaCount
-
-                playlistList.add(playlistData)
+                artistList.add(tempData)
             }
 
             // Close cursor to avoid memory leaks.
             cursor?.close()
             // After finish the "query", go back to the "main" thread(You can only call flutter
             // inside the main thread).
-            return@withContext playlistList
+            return@withContext artistList
         }
 }
 
 //Extras:
 
-//I/OnPlaylistCursor[All/Audio]: [
-// _data
+//I/OnArtistCursor[All/Audio]: [
 // _id
-// date_added
-// date_modified
-// name
+// artist
+// artist_key
+// number_of_albums
+// number_of_tracks
 // ]
