+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/methods/queries/GenresQuery.kt
@@ -1,45 +1,45 @@
-package com.lucasjosino.on_audio_query.query
+package com.lucasjosino.on_audio_query.methods
 
 import android.content.ContentResolver
 import android.content.Context
 import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.controller.PermissionController
-import com.lucasjosino.on_audio_query.query.helper.QueryHelper
-import com.lucasjosino.on_audio_query.types.checkArtistsUriType
-import com.lucasjosino.on_audio_query.types.sorttypes.checkArtistSortType
-import com.lucasjosino.on_audio_query.utils.artistProjection
+import com.lucasjosino.on_audio_query.controllers.PermissionController
+import com.lucasjosino.on_audio_query.methods.helper.QueryHelper
+import com.lucasjosino.on_audio_query.types.checkGenresUriType
+import com.lucasjosino.on_audio_query.types.sorttypes.checkGenreSortType
+import com.lucasjosino.on_audio_query.utils.genreProjection
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-/** ArtistsQuery */
-class ArtistsQuery : ViewModel() {
+/** GenresQuery */
+class GenresQuery : ViewModel() {
 
-    //Main parameters
+    // Main parameters.
     private val helper = QueryHelper()
 
     // None of this methods can be null.
     private lateinit var uri: Uri
-    private lateinit var resolver: ContentResolver
     private lateinit var sortType: String
+    private lateinit var resolver: ContentResolver
 
     /**
-     * Method to "query" all artists.
+     * Method to "query" all genres.
      *
      * Parameters:
      *   * [context]
      *   * [result]
      *   * [call]
      */
-    fun queryArtists(context: Context, result: MethodChannel.Result, call: MethodCall) {
+    fun queryGenres(context: Context, result: MethodChannel.Result, call: MethodCall) {
         resolver = context.contentResolver
 
-        // Sort: Type and Order
-        sortType = checkArtistSortType(
+        // Sort: Type and Order.
+        sortType = checkGenreSortType(
             call.argument<Int>("sortType"),
             call.argument<Int>("orderType")!!,
             call.argument<Boolean>("ignoreCase")!!
@@ -47,59 +47,62 @@ class ArtistsQuery : ViewModel() {
         // Check uri:
         //   * [0]: External.
         //   * [1]: Internal.
-        uri = checkArtistsUriType(call.argument<Int>("uri")!!)
+        uri = checkGenresUriType(call.argument<Int>("uri")!!)
 
         // Query everything in background for a better performance.
         viewModelScope.launch {
             // Request permission status from the main method.
             val hasPermission = PermissionController().permissionStatus(context)
             // Empty list.
-            var resultArtistList = ArrayList<MutableMap<String, Any?>>()
+            var resultGenreList = ArrayList<MutableMap<String, Any?>>()
 
             // We cannot "query" without permission so, just return a empty list.
             if (hasPermission) {
                 // Start querying
-                resultArtistList = loadArtists()
+                resultGenreList = loadGenres()
             }
 
             //Flutter UI will start, but, information still loading
-            result.success(resultArtistList)
+            result.success(resultGenreList)
         }
     }
 
-    // Loading in Background
-    private suspend fun loadArtists(): ArrayList<MutableMap<String, Any?>> =
+    //Loading in Background
+    private suspend fun loadGenres(): ArrayList<MutableMap<String, Any?>> =
         withContext(Dispatchers.IO) {
             // Setup the cursor with [uri], [projection] and [sortType].
-            val cursor = resolver.query(uri, artistProjection, null, null, sortType)
+            val cursor = resolver.query(uri, genreProjection, null, null, sortType)
             // Empty list.
-            val artistList: ArrayList<MutableMap<String, Any?>> = ArrayList()
+            val genreList: ArrayList<MutableMap<String, Any?>> = ArrayList()
 
-            // For each item(artist) inside this "cursor", take one and "format"
+            // For each item(genre) inside this "cursor", take one and "format"
             // into a [Map<String, dynamic>].
             while (cursor != null && cursor.moveToNext()) {
-                val tempData: MutableMap<String, Any?> = HashMap()
-                for (artistMedia in cursor.columnNames) {
-                    tempData[artistMedia] = helper.loadArtistItem(artistMedia, cursor)
+                val genreData: MutableMap<String, Any?> = HashMap()
+                for (genreMedia in cursor.columnNames) {
+                    genreData[genreMedia] = helper.loadGenreItem(genreMedia, cursor)
                 }
 
-                artistList.add(tempData)
+                // Count and add the number of songs for every genre.
+                val mediaCount = helper.getMediaCount(0, genreData["_id"].toString(), resolver)
+                genreData["num_of_songs"] = mediaCount
+
+                if (genreData["name"] != null && genreData["_id"] != 0) {
+                    genreList.add(genreData)
+                }
             }
 
             // Close cursor to avoid memory leaks.
             cursor?.close()
             // After finish the "query", go back to the "main" thread(You can only call flutter
             // inside the main thread).
-            return@withContext artistList
+            return@withContext genreList
         }
 }
 
 //Extras:
 
-//I/OnArtistCursor[All/Audio]: [
+//I/OnGenreCursor[All/Audio]: [
 // _id
-// artist
-// artist_key
-// number_of_albums
-// number_of_tracks
+// name
 // ]
