+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/methods/queries/PlaylistsQuery.kt
@@ -1,41 +1,41 @@
-package com.lucasjosino.on_audio_query.query
+package com.lucasjosino.on_audio_query.methods
 
 import android.content.ContentResolver
 import android.content.Context
 import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.controller.PermissionController
-import com.lucasjosino.on_audio_query.query.helper.QueryHelper
-import com.lucasjosino.on_audio_query.types.checkGenresUriType
+import com.lucasjosino.on_audio_query.controllers.PermissionController
+import com.lucasjosino.on_audio_query.methods.helper.QueryHelper
+import com.lucasjosino.on_audio_query.types.checkPlaylistsUriType
 import com.lucasjosino.on_audio_query.types.sorttypes.checkGenreSortType
-import com.lucasjosino.on_audio_query.utils.genreProjection
+import com.lucasjosino.on_audio_query.utils.playlistProjection
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-/** GenresQuery */
-class GenresQuery : ViewModel() {
+/** PlaylistsQuery */
+class PlaylistsQuery : ViewModel() {
 
-    // Main parameters.
+    //Main parameters
     private val helper = QueryHelper()
 
     // None of this methods can be null.
     private lateinit var uri: Uri
-    private lateinit var sortType: String
     private lateinit var resolver: ContentResolver
+    private lateinit var sortType: String
 
     /**
-     * Method to "query" all genres.
+     * Method to "query" all playlists.
      *
      * Parameters:
      *   * [context]
      *   * [result]
      *   * [call]
      */
-    fun queryGenres(context: Context, result: MethodChannel.Result, call: MethodCall) {
+    fun queryPlaylists(context: Context, result: MethodChannel.Result, call: MethodCall) {
         resolver = context.contentResolver
 
         // Sort: Type and Order.
@@ -47,62 +47,63 @@ class GenresQuery : ViewModel() {
         // Check uri:
         //   * [0]: External.
         //   * [1]: Internal.
-        uri = checkGenresUriType(call.argument<Int>("uri")!!)
+        uri = checkPlaylistsUriType(call.argument<Int>("uri")!!)
 
         // Query everything in background for a better performance.
         viewModelScope.launch {
             // Request permission status from the main method.
             val hasPermission = PermissionController().permissionStatus(context)
             // Empty list.
-            var resultGenreList = ArrayList<MutableMap<String, Any?>>()
+            var resultPlaylistList = ArrayList<MutableMap<String, Any?>>()
 
             // We cannot "query" without permission so, just return a empty list.
             if (hasPermission) {
                 // Start querying
-                resultGenreList = loadGenres()
+                resultPlaylistList = loadPlaylists()
             }
 
             //Flutter UI will start, but, information still loading
-            result.success(resultGenreList)
+            result.success(resultPlaylistList)
         }
     }
 
     //Loading in Background
-    private suspend fun loadGenres(): ArrayList<MutableMap<String, Any?>> =
+    private suspend fun loadPlaylists(): ArrayList<MutableMap<String, Any?>> =
         withContext(Dispatchers.IO) {
-            // Setup the cursor with [uri], [projection] and [sortType].
-            val cursor = resolver.query(uri, genreProjection, null, null, sortType)
+            // Setup the cursor with [uri] and [projection].
+            val cursor = resolver.query(uri, playlistProjection, null, null, null)
             // Empty list.
-            val genreList: ArrayList<MutableMap<String, Any?>> = ArrayList()
+            val playlistList: ArrayList<MutableMap<String, Any?>> = ArrayList()
 
-            // For each item(genre) inside this "cursor", take one and "format"
+            // For each item(playlist) inside this "cursor", take one and "format"
             // into a [Map<String, dynamic>].
             while (cursor != null && cursor.moveToNext()) {
-                val genreData: MutableMap<String, Any?> = HashMap()
-                for (genreMedia in cursor.columnNames) {
-                    genreData[genreMedia] = helper.loadGenreItem(genreMedia, cursor)
+                val playlistData: MutableMap<String, Any?> = HashMap()
+                for (playlistMedia in cursor.columnNames) {
+                    playlistData[playlistMedia] = helper.loadPlaylistItem(playlistMedia, cursor)
                 }
 
-                // Count and add the number of songs for every genre.
-                val mediaCount = helper.getMediaCount(0, genreData["_id"].toString(), resolver)
-                genreData["num_of_songs"] = mediaCount
+                // Count and add the number of songs for every playlist.
+                val mediaCount = helper.getMediaCount(1, playlistData["_id"].toString(), resolver)
+                playlistData["num_of_songs"] = mediaCount
 
-                if (genreData["name"] != null && genreData["_id"] != 0) {
-                    genreList.add(genreData)
-                }
+                playlistList.add(playlistData)
             }
 
             // Close cursor to avoid memory leaks.
             cursor?.close()
             // After finish the "query", go back to the "main" thread(You can only call flutter
             // inside the main thread).
-            return@withContext genreList
+            return@withContext playlistList
         }
 }
 
 //Extras:
 
-//I/OnGenreCursor[All/Audio]: [
+//I/OnPlaylistCursor[All/Audio]: [
+// _data
 // _id
+// date_added
+// date_modified
 // name
 // ]
